# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Syba-Backend
# during dev, should point to your local machine
host: localhost:7000
...
basePath: /api
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /batchconfigs/{BOID}/reload:
    x-swagger-router-controller: batches
    post:
      description: Reloads a batchconfig completly, recalculates all runs and all kennzahlen
      operationId: reloadBatchConfig
      parameters:
        - name: BOID
          in: path
          description: BOID der Batchconfig, die reloaded werden soll
          required: true
          type: string
      responses:
        "200":
          description: Success
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /batchconfigs/{ITSBATCHCONFIG}/kennzahlvalues/{ITSKENNZAHLCONFIG}:
    x-swagger-router-controller: kennzahlvalues
    get: 
      description: Returns a kennzahl history for one batchconfig/kennzahlconfig
      operationId: getKennzahlWithHistory
      parameters: 
        - name: ITSBATCHCONFIG
          in: path
          description: boid of batchconfig to fetch KennzahlValues for
          required: true
          type: string
        - name: ITSKENNZAHLCONFIG
          in: path
          description: boid of kennzahlConfig to fetch KennzahlValues for
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            type: array
            items: 
              # a pointer to a definition
              $ref: "#/definitions/KennzahlValue"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /batchconfigs/{itsBatchConfig}/run/{itsBatchRun}/kennzahlen:
    x-swagger-router-controller: kennzahlvalues
    get:
      description: Returns all kennzahlValues for a batchconfig
      operationId: getKennzahlenForRun
      parameters:
        - name: itsBatchConfig
          in: path
          description: boid of batchconfig to fetch KennzahlValues for
          required: true
          type: string
        - name: itsBatchRun
          in: path
          description: boid of batchconfig to fetch KennzahlValues for
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              # a pointer to a definition
              $ref: "#/definitions/RunKennzahl"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  KennzahlValue:
    type: object
    properties:
      boid:
        type: string
      itsBatchRun:
        type: string
      itsKennzahlConfig:
        type: string
      itsBatchconfig:
        type: string
      itsSyriusBatch:
        type: string
      numbervalue:
        type: string
      timestampValue:
        type: string
        format: date-time
  RunKennzahl:
    type: object
    properties:
      ITSBATCHRUN:
        type: string
      ITSKENNZAHLCONFIG:
        type: string
      ITSBATCHCONFIG:
        type: string
      ITSKENNZAHL:
        type: string
      NAME:
        type: string
      VALUE:
        type: string
      AVG:
        type: string
      STDDEV:
        type: string
      MIN:
        type: string
      MAX:
        type: string
...